Fixing condition for setting rootfs_dest_backing.
The fix for the opt_link check in d5ed942 broke the code as it inverted the
meaning. Previously it ran the code iff opt_link was empty, as [ -n ] is
true. In the case opt_link was not empty, the code evaluated to [ -n foo],
which is invalid syntax and thus false. This, however, was almost the expected
behaviour, as rootfs_dest_backing is needed for opt_union and (in read-only
mode) for opt_link.