Fix removable device flag handling for live-medium
In the past most removable USB flash devices had the removable flag set.
This was wrong as that flag means that the device contains a removable
medium. But the flash chips in most of these devices are NOT removable.

This patch fixes this by assuming that all USB devices are removable.
There is NO way to really know, so this is the most sensible thing to
assume. To make this clear, the "removable-usb" parameter is deprecated
and now called just "usb".

This patch uses the same logic as used by udisks to determine which
devices should be considered removable. Beside the removable flag for
devices containing removable media it uses some heuristics for which
buses contain removable devices. Devices on USB, Firewire and MMC buses
are considered removable.